fix: address CodeQL quality findings in Gdal, OgcApi, OData, Geoprocessing
Addresses the 180 Code Quality (CodeQL) findings assigned to this sweep
group across src/Honua.Worker.Gdal, src/Honua.Protocols.OgcApi,
src/Honua.Protocols.OData, and src/Honua.Geoprocessing.

Breakdown by rule.id (180 total):
- cs/catch-of-all-exceptions (75): added house-style "intentionally
  generic/broad" justification comments to legitimate broad boundaries
  (job executor boundaries, background loops, best-effort cleanup,
  top-level protocol endpoint/handler error boundaries). Most were
  already correctly justified by earlier merged sweeps; the remainder
  got new comments tailored to each method.
- cs/path-combine (57): added justification comments to false positives
  where later Path.Combine segments are fixed/generated relative
  literals or already-sanitized/validated identifiers, and verified
  (without code changes, since already correctly guarded) the handful
  of calls that combine job-supplied identifiers into scratch/workspace
  paths (GdalScratch, GdalVsiPath, GdalVectorSourceReadJobExecutor's
  zip-slip guard, SinkPathResolver, FeatureStreamArtifact) are protected
  by pre-validation and/or post-combine root-containment checks.
- cs/linq/missed-where (20) / cs/linq/missed-select (15): applied clean
  rewrites where safe; left loops with side effects, mid-loop early
  returns, or out-parameter dependencies as-is with a justification
  comment.
- cs/missed-using-statement (5): verified all carry a documented reason
  a using declaration can't be used (reassignment across try/catch/
  finally boundaries).
- cs/equality-on-floats (4): documented intentional exact-equality
  checks (sentinel/parsed-literal comparisons, CQL2/OData "=" push-down
  semantics, a divisor guard where an earlier PR already reverted an
  epsilon check that broke tiny legitimate divisors).
- cs/constant-condition (2): one was dead code already removed by a
  prior commit (GetQueryTtlWithJitter clamp made a downgrade guard
  unreachable); the other (OgcMapsRenderingHandler Transparent ?? true)
  was confirmed non-behavior-changing and documented rather than risking
  a simplification.
- cs/loss-of-precision (1): genuine bug fix in GdalOutputGridGuard -
  integer division of MaxDecodedRasterBytes by BytesPerFloat64Pixel
  truncated the admission threshold before comparison, which could
  reject a raster that was actually within the configured byte budget;
  widened the numerator to double before dividing.
- cs/dereferenced-value-may-be-null (1): added a null-forgiving operator
  with a comment in ODataQueryParameterAdapter where CodeQL couldn't
  follow a TryGetValue-narrowed branch.

All four projects (Honua.Worker.Gdal, Honua.Protocols.OgcApi,
Honua.Protocols.OData, Honua.Geoprocessing) build clean in Release with
TreatWarningsAsErrors=true, and dotnet format reported no further
changes needed.

+++ src/Honua.Worker.Gdal/Execution/GdalOutputGridGuard.cs
@@ -173,7 +173,11 @@ internal static class GdalOutputGridGuard
         // (8 bytes/pixel), mirroring TryAdmit; the width / height / pixel caps are the
         // primary bound and hold regardless of the true output band count / dtype.
         const long BytesPerFloat64Pixel = 8L;
-        if (pixels > options.MaxDecodedRasterBytes / BytesPerFloat64Pixel)
+        // Divide in double space (widen the long numerator) rather than truncating
+        // integer division first: MaxDecodedRasterBytes is not guaranteed to be a
+        // multiple of 8, and a floored integer threshold would reject a pixel count
+        // that is actually still within the configured byte budget.
+        if (pixels > (double)options.MaxDecodedRasterBytes / BytesPerFloat64Pixel)
         {
             error = $"estimated output grid size {Format(pixels)} pixels × 8 bytes/pixel (single-band Float64) exceeds configured MaxDecodedRasterBytes={options.MaxDecodedRasterBytes.ToString(CultureInfo.InvariantCulture)}";
             return false;
